requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.2
django-debug-toolbar==5.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycodestyle==2.13.0
PyJWT==2.9.0
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1